Deduplicate US tax module and remove dead code
Deduplicated:
- src/utils/constants.ts now re-exports TAX_BRACKETS_MFJ/SINGLE,
  STANDARD_DEDUCTION_MFJ/SINGLE, CAPITAL_GAINS_BRACKETS_MFJ/SINGLE,
  RMD_START_AGE, RMD_TABLE, and getRMDDivisor from
  src/countries/usa/constants.ts (now the canonical source, where
  getRMDDivisor was added). App-level items (CHART_COLORS,
  DEFAULT_PROFILE, DEFAULT_ASSUMPTIONS, DEFAULT_INCOME_STREAMS) remain
  defined locally.
- src/utils/taxes.ts now delegates getTaxBrackets, getStandardDeduction,
  getCapitalGainsBrackets, calculateFederalIncomeTax,
  calculateCapitalGainsTax, and calculateTotalFederalTax to
  src/countries/usa/taxes.ts via thin wrappers that preserve the
  strict FilingStatus signatures. calculateStateTax stays as-is (not
  duplicated in usa/taxes.ts).

Removed (verified unreferenced anywhere in src/, including
src/tests/calculations.test.ts):
- clearAllData in src/hooks/useLocalStorage.ts (also referenced the
  wrong localStorage keys)
- getWithdrawalToFillBracket, getMarginalTaxRate, getEffectiveTaxRate
  in src/utils/taxes.ts
- getBalanceAtAge, calculateTotalContributions in
  src/utils/projections.ts

Behavior-neutral: all 166 tests pass unchanged, build and lint clean.
