+++ src/utils/projections.ts
@@ -118,45 +118,3 @@ export function calculateAccumulation(
     breakdownByGroup,
   };
 }
-
-/**
- * Get the balance of an account at a specific age
- */
-export function getBalanceAtAge(
-  result: AccumulationResult,
-  accountId: string,
-  age: number
-): number {
-  const yearData = result.yearlyBalances.find(y => y.age === age);
-  if (!yearData) return 0;
-  return yearData.balances[accountId] || 0;
-}
-
-/**
- * Calculate total contributions made over accumulation phase
- */
-export function calculateTotalContributions(
-  accounts: Account[],
-  profile: Profile
-): Record<string, number> {
-  const yearsToRetirement = profile.retirementAge - profile.currentAge;
-  const totals: Record<string, number> = {};
-
-  accounts.forEach(account => {
-    let totalContribution = 0;
-    let yearlyContribution = account.annualContribution;
-
-    for (let i = 0; i < yearsToRetirement; i++) {
-      const employerMatch = calculateEmployerMatch({
-        ...account,
-        annualContribution: yearlyContribution,
-      });
-      totalContribution += yearlyContribution + employerMatch;
-      yearlyContribution *= (1 + account.contributionGrowthRate);
-    }
-
-    totals[account.id] = totalContribution;
-  });
-
-  return totals;
-}
